+++ application/modules/guestbook/controllers/admin/Index.php
@@ -40,6 +40,27 @@ class Index extends \Ilch\Controller\Admin
     public function indexAction()
     {
         $guestbookMapper = new GuestbookMapper();
+        
+        if ($this->getRequest()->getPost('check_entries')) {
+            if ($this->getRequest()->getPost('action') == 'delete') {
+                foreach($this->getRequest()->getPost('check_entries') as $entryId) {
+                    $guestbookMapper->delete($entryId);
+                }
+                
+                $this->redirect(array('action' => 'index'));
+            }
+
+            if ($this->getRequest()->getPost('action') == 'setfree') {
+                foreach($this->getRequest()->getPost('check_entries') as $entryId) {
+                    $model = new \Guestbook\Models\Entry();
+                    $model->setId($entryId);
+                    $model->setFree(1);
+                    $guestbookMapper->save($model);
+                }
+
+                $this->redirect(array('action' => 'index'));
+            }
+        }
 
         if ($this->getRequest()->getParam('showsetfree')) {
             $entries = $guestbookMapper->getEntries(array('setfree' => 0));
@@ -54,9 +75,12 @@ class Index extends \Ilch\Controller\Admin
     public function delAction()
     {
         $guestbookMapper = new GuestbookMapper();
-        $guestbookMapper->delete($this->getRequest()->getParam('id'));
-        $this->addMessage('deleteSuccess');
         
+        if($this->getRequest()->isSecure()) {
+            $guestbookMapper->delete($this->getRequest()->getParam('id'));
+            $this->addMessage('deleteSuccess');
+        }
+
         if ($this->getRequest()->getParam('showsetfree')) {
             $this->redirect(array('action' => 'index', 'showsetfree' => 1));
         } else {
